Add 'concatenate' option
In some situations, we'd like to output the compiled templates as
individual files. For example, suppose you have a dependency graph where
there are many views, and the templates corresponding to each view are
discoverable via the graph.

In this case, it makes sense to keep all templates as individual files,
and allow the build optimizer to concatenate the templates along with
the individual view's code, so that views can be loaded lazily.